Separate app and server initialisation
